Pulse the LED to indicate direction of productivity change.

+++ firmware/src/rescuetime.h
@@ -12,7 +12,7 @@ class RescueTime
 public:
     RescueTime(String);
 
-    int getProductivityScore();
+    float getProductivityScore();
 
 private:
     String api_key;
